move abstract to log clearner lag integration test
