+++ storage/src/test/java/org/apache/kafka/storage/internals/log/LogCleanerLagIntegrationTest.java
@@ -18,21 +18,37 @@ package org.apache.kafka.storage.internals.log;
 
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.compress.Compression;
+import org.apache.kafka.common.config.TopicConfig;
 import org.apache.kafka.common.record.internal.CompressionType;
+import org.apache.kafka.common.record.internal.MemoryRecords;
 import org.apache.kafka.common.record.internal.Record;
+import org.apache.kafka.common.record.internal.RecordBatch;
+import org.apache.kafka.common.record.internal.RecordVersion;
+import org.apache.kafka.common.utils.Utils;
+import org.apache.kafka.coordinator.transaction.TransactionLogConfig;
 import org.apache.kafka.server.util.MockTime;
 
-import org.junit.jupiter.api.Tag;
+import org.apache.kafka.storage.log.metrics.BrokerTopicStats;
+import org.apache.kafka.test.TestUtils;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
 import java.io.IOException;
-import java.util.Arrays;
+import java.nio.file.Files;
+import java.time.Duration;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
+import java.util.Properties;
+import java.util.Random;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -40,33 +56,34 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 /**
  * This is an integration test that tests the fully integrated log cleaner
  */
-@Tag("integration")
-public class LogCleanerLagIntegrationTest extends AbstractLogCleanerIntegrationTest {
-
+public class LogCleanerLagIntegrationTest {
     private static final Logger log = LoggerFactory.getLogger(LogCleanerLagIntegrationTest.class);
 
-    private static final int MS_PER_HOUR = 60 * 60 * 1000;
-    private static final int MIN_COMPACTION_LAG = MS_PER_HOUR; // 1 hour
+    protected LogCleaner cleaner;
+    protected final File logDir = TestUtils.tempDirectory();
 
-    static {
-        // compactionLag must be divisible by 2 for this test
-        assertTrue(MIN_COMPACTION_LAG % 2 == 0, "compactionLag must be divisible by 2 for this test");
-    }
-
-    private final MockTime time = new MockTime(1400000000000L, 1000L);  // Tue May 13 16:53:20 UTC 2014
+    private final List<UnifiedLog> logs = new ArrayList<>();
+    private static final int DEFAULT_MAX_MESSAGE_SIZE = 128;
+    private static final int DEFAULT_DELETE_DELAY = 1000;
+    private static final int DEFAULT_SEGMENT_SIZE = 2048;
+    private static final long DEFAULT_MIN_COMPACTION_LAG_MS = 0L;
+    private static final long DEFAULT_MAX_COMPACTION_LAG_MS = Long.MAX_VALUE;
+    private static final long MIN_COMPACTION_LAG = Duration.ofHours(1).toMillis();
     private static final long CLEANER_BACKOFF_MS = 200L;
+    private static final float DEFAULT_MIN_CLEANABLE_DIRTY_RATIO = 0.0F;
     private static final int SEGMENT_SIZE = 512;
 
-    private static final List<TopicPartition> TOPIC_PARTITIONS = Arrays.asList(
+    private int counter = 0;
+
+    private final MockTime time = new MockTime(1400000000000L, 1000L);  // Tue May 13 16:53:20 UTC 2014
+    private static final List<TopicPartition> TOPIC_PARTITIONS = List.of(
         new TopicPartition("log", 0),
         new TopicPartition("log", 1),
         new TopicPartition("log", 2)
     );
 
-    @Override
-    protected MockTime time() {
-        return time;
-    }
+    public record KeyValueOffset(int key, String value, long firstOffset) { }
+    public record ValueAndRecords(String value, MemoryRecords records) { }
 
     @ParameterizedTest
     @EnumSource(CompressionType.class)
@@ -165,5 +182,178 @@ public class LogCleanerLagIntegrationTest extends AbstractLogCleanerIntegrationT
             }
         }
         return result;
+    }
+
+    private Properties logConfigProperties(Properties propertyOverrides,
+                                           int maxMessageSize,
+                                           float minCleanableDirtyRatio,
+                                           long minCompactionLagMs,
+                                           int deleteDelay,
+                                           int segmentSize,
+                                           long maxCompactionLagMs) {
+        Properties props = new Properties();
+        props.put(TopicConfig.MAX_MESSAGE_BYTES_CONFIG, maxMessageSize);
+        props.put(LogConfig.INTERNAL_SEGMENT_BYTES_CONFIG, segmentSize);
+        props.put(TopicConfig.SEGMENT_INDEX_BYTES_CONFIG, 100 * 1024);
+        props.put(TopicConfig.FILE_DELETE_DELAY_MS_CONFIG, deleteDelay);
+        props.put(TopicConfig.CLEANUP_POLICY_CONFIG, TopicConfig.CLEANUP_POLICY_COMPACT);
+        props.put(TopicConfig.MIN_CLEANABLE_DIRTY_RATIO_CONFIG, minCleanableDirtyRatio);
+        props.put(TopicConfig.MIN_COMPACTION_LAG_MS_CONFIG, minCompactionLagMs);
+        props.put(TopicConfig.MAX_COMPACTION_LAG_MS_CONFIG, maxCompactionLagMs);
+        props.putAll(propertyOverrides);
+        return props;
+    }
+
+    private Properties logConfigProperties(int maxMessageSize) {
+        return logConfigProperties(new Properties(), maxMessageSize,
+            DEFAULT_MIN_CLEANABLE_DIRTY_RATIO, DEFAULT_MIN_COMPACTION_LAG_MS,
+            DEFAULT_DELETE_DELAY, DEFAULT_SEGMENT_SIZE, DEFAULT_MAX_COMPACTION_LAG_MS);
+    }
+
+    private LogCleaner makeCleaner(Iterable<TopicPartition> partitions,
+                                   float minCleanableDirtyRatio,
+                                   int numThreads,
+                                   long backoffMs,
+                                   int maxMessageSize,
+                                   long minCompactionLagMs,
+                                   int deleteDelay,
+                                   int segmentSize,
+                                   long maxCompactionLagMs,
+                                   Integer cleanerIoBufferSize,
+                                   Properties propertyOverrides) throws IOException {
+
+        ConcurrentMap<TopicPartition, UnifiedLog> logMap = new ConcurrentHashMap<>();
+        for (TopicPartition partition : partitions) {
+            File dir = new File(logDir, partition.topic() + "-" + partition.partition());
+            Files.createDirectories(dir.toPath());
+
+            Properties props = logConfigProperties(propertyOverrides,
+                maxMessageSize,
+                minCleanableDirtyRatio,
+                minCompactionLagMs,
+                deleteDelay,
+                segmentSize,
+                maxCompactionLagMs);
+            LogConfig logConfig = new LogConfig(props);
+
+            UnifiedLog log = UnifiedLog.create(
+                dir,
+                logConfig,
+                0L,
+                0L,
+                time.scheduler,
+                new BrokerTopicStats(),
+                time,
+                5 * 60 * 1000,
+                new ProducerStateManagerConfig(TransactionLogConfig.PRODUCER_ID_EXPIRATION_MS_DEFAULT, false),
+                TransactionLogConfig.PRODUCER_ID_EXPIRATION_CHECK_INTERVAL_MS_DEFAULT,
+                new LogDirFailureChannel(10),
+                true,
+                Optional.empty());
+            logMap.put(partition, log);
+            logs.add(log);
+        }
+
+        int ioBufferSize = cleanerIoBufferSize != null ? cleanerIoBufferSize : maxMessageSize / 2;
+        CleanerConfig cleanerConfig = new CleanerConfig(
+            numThreads,
+            4 * 1024 * 1024L,
+            0.9,
+            ioBufferSize,
+            maxMessageSize,
+            Double.MAX_VALUE,
+            backoffMs,
+            true);
+
+        return new LogCleaner(cleanerConfig,
+            List.of(logDir),
+            logMap,
+            new LogDirFailureChannel(1),
+            time);
+    }
+
+    private LogCleaner makeCleaner(Iterable<TopicPartition> partitions,
+                                   long backoffMs,
+                                   long minCompactionLagMs,
+                                   int segmentSize) throws IOException {
+        return makeCleaner(partitions,
+            DEFAULT_MIN_CLEANABLE_DIRTY_RATIO,
+            1,
+            backoffMs,
+            DEFAULT_MAX_MESSAGE_SIZE,
+            minCompactionLagMs,
+            DEFAULT_DELETE_DELAY,
+            segmentSize,
+            DEFAULT_MAX_COMPACTION_LAG_MS,
+            null,
+            new Properties());
+    }
+
+    private int counter() {
+        return counter;
+    }
+
+    private void incCounter() {
+        counter++;
+    }
+
+    private List<KeyValueOffset> writeDups(int numKeys, int numDups, UnifiedLog log, Compression codec,
+                                           int startKey, byte magicValue) throws IOException {
+        List<KeyValueOffset> results = new ArrayList<>();
+        for (int i = 0; i < numDups; i++) {
+            for (int key = startKey; key < startKey + numKeys; key++) {
+                String value = String.valueOf(counter());
+                MemoryRecords records = LogTestUtils.singletonRecords(
+                    value.getBytes(),
+                    codec,
+                    String.valueOf(key).getBytes(),
+                    RecordBatch.NO_TIMESTAMP,
+                    magicValue);
+                LogAppendInfo appendInfo = log.appendAsLeaderWithRecordVersion(
+                    records, 0, RecordVersion.lookup(magicValue));
+                // move LSO forward to increase compaction bound
+                log.updateHighWatermark(log.logEndOffset());
+                results.add(new KeyValueOffset(key, value, appendInfo.firstOffset()));
+                incCounter();
+            }
+        }
+        return results;
+    }
+
+    private List<KeyValueOffset> writeDups(int numKeys, int numDups, UnifiedLog log, Compression codec) throws IOException {
+        return writeDups(numKeys, numDups, log, codec, 0, RecordBatch.CURRENT_MAGIC_VALUE);
+    }
+
+    private ValueAndRecords createLargeSingleMessageSet(int key, byte messageFormatVersion, Compression codec) {
+        Random random = new Random(0);
+        StringBuilder sb = new StringBuilder(128);
+        for (int i = 0; i < 128; i++) {
+            sb.append((char) ('a' + random.nextInt(26)));
+        }
+        String value = sb.toString();
+        MemoryRecords records = LogTestUtils.singletonRecords(
+            value.getBytes(),
+            codec,
+            String.valueOf(key).getBytes(),
+            RecordBatch.NO_TIMESTAMP,
+            messageFormatVersion);
+        return new ValueAndRecords(value, records);
+    }
+
+    private void closeLog(UnifiedLog log) throws IOException {
+        log.close();
+        logs.remove(log);
+    }
+
+    @AfterEach
+    public void teardown() throws IOException, InterruptedException {
+        if (cleaner != null) {
+            cleaner.shutdown();
+        }
+        time.scheduler.shutdown();
+        for (UnifiedLog log : logs) {
+            log.close();
+        }
+        Utils.delete(logDir);
     }
 }
